Command split: Implement the "ruby" command in a separate class
